% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN1toKmaxPredict.R
\name{NN1toKmaxPredict}
\alias{NN1toKmaxPredict}
\title{NN1toKmaxPredict Algo}
\usage{
NN1toKmaxPredict(X.mat, Y.vec, testX.mat, max_neighbors)
}
\arguments{
\item{X.mat}{numeric input feature matrix [n x p]}

\item{Y.vec}{numeric input label vetor [n], either all 0/1 for bninary classification or other real numbers for regression, it doesn't support others}

\item{testX.mat}{numeric test matrix}

\item{max_neighbors}{scalar integer, max number of neighbors}
}
\value{
numeric vector size of n x max_neighbors, predictions from 1 to max_neighbors for all rows of testX.mat
}
\description{
Coding Project 1 R function for NN1toKmaxPredict Algo
}
\examples{
data(zip.train, package = "ElemStatLearn")
io1 <- which(zip.train[,1] \%in\% c(0,1))
train.i <- io1[1:5]
test.i <- io1[6:10]
X <- zip.train[train.i, -1]
Y <- zip.train[train.i, 1]
testX <- zip.train[test.i, -1]
max_neighbors <- 3
NN1toKmaxPredict(X,Y,testX, 3)

data(prostate, package = "ElemStatLearn")
io1 <- which(zip.train[,1] \%in\% c(0,1))
train.i <- io1[1:5]
test.i <- io1[6:10]
X <- zip.train[train.i, -1]
Y <- zip.train[train.i, 1]
testX <- zip.train[test.i, -1]
max_neighbors <- 3
NN1toKmaxPredict(X,Y,testX, 3)
}
